+++ HISTORY.md
@@ -3,10 +3,14 @@
 ## in development:
 
 * **Notable Changes:**
+    * optimized special cases for COMBINE (sum, product, min, max) by up to 34 x
     * custom block label parts inside the prototype (in the block editor) are now displayed the same as in block instances
     * variadic ring inputs are now arranged vertically (e.g. the reporter rings in PIPE)
     * changed zebra-coloring for yellow custom block prototypes (in the block editor) so the hat block changes the shade, not the prototype
     * improved layout and rendering of (+) buttons in custom block prototypes
+
+### 2021-03-02
+* threads: optimized special cases for COMBINE (sum, product, min, max) by up to 34 x
 
 ### 2021-03-01
 * byob: improved layout and rendering of (+) buttons in custom block prototypes